move examples into subdirectory
