+++ examples/full/Dockerfile
