+++ examples/full/docker-compose.yml
