+++ examples/full/package.json
@@ -7,7 +7,7 @@
   "dependencies": {
     "aws-sdk": "^2.77.0",
     "bluebird": "^3.5.0",
-    "bunyan": "^1.8.12",
+    "bunyan": "^1.8.15",
     "generic-pool": "^3.7.2",
     "lazy-object": "^1.0.1",
     "lodash.difference": "^4.5.0",
