+++ examples/mp/README.md
@@ -1,0 +1,6 @@
+multi-process example
+
+```bash
+yarn
+yarn ts-node index.ts | ./node_modules/.bin/bunyan
+```
