+++ examples/mp/index.ts
