+++ examples/mp/logger.ts
