+++ examples/mp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test-kafka",
+  "name": "test-mp",
   "version": "1.0.0",
   "description": "test steveo",
   "main": "index.js",
