+++ examples/mp/tasks/spam_task.ts
